rename ErrorBoundary
