+++ src/ErrorBoundary/index.ts
@@ -1,1 +1,1 @@
-export * from './errorBoundary'
+export * from './ErrorBoundary'
